evga_automated_order/requirements.txt
selenium==3.14.0
requests==2.12.4
pyyaml==3.12